requirements.txt
nextcord
setuptools
yt-dlp
PyNaCl
python-dotenv
